basic lua sandboxing
